+++ lib/netzke/core/component_config.rb
@@ -1,11 +1,12 @@
 module Netzke::Core
-  class ComponentConfig < ActiveSupport::OrderedOptions
-    def initialize(name, parent)
+  class ComponentConfig < DslConfigBase
+    def initialize(name, component)
       self.name = name.to_s
     end
 
     def set_defaults!
       self.item_id ||= name # default item_id
+      self.client_config ||= {}
       self.klass ||= name.camelize.constantize # default klass
     end
   end