+++ lib/netzke/core/dynamic_assets.rb
@@ -32,11 +32,7 @@ module Netzke
         end
 
         def strip_js_comments(js_string)
-          if defined?(::Rails) && !::Rails.env.development? && compressor = ::Rails.application.assets.js_compressor
-            compressor.processor.call(nil, js_string)
-          else
-            js_string
-          end
+          js_string
         end
 
       private
@@ -48,7 +44,7 @@ module Netzke
           res << %{Ext.ns('Netzke');}
           res << %{Ext.ns('Netzke.core');}
           res << %{Netzke.RelativeUrlRoot = '#{ActionController::Base.config.relative_url_root}';}
-          res << %{Netzke.ControllerUrl = '#{ActionController::Base.config.relative_url_root}#{Rails.application.routes.url_helpers.netzke_path('')}';}
+          res << %{Netzke.ControllerUrl = '#{ActionController::Base.config.relative_url_root}#{Rails.application.routes.url_helpers.netzke_path}/';}
           res << %{Netzke.RelativeExtUrl = '#{ActionController::Base.config.relative_url_root}#{Netzke::Core.ext_uri}';}
 
           res << %{Netzke.core.directMaxRetries = #{Netzke::Core.js_direct_max_retries};}