+++ lib/netzke/core/railz/action_view_ext/ext.rb
@@ -8,10 +8,10 @@ module Netzke
 
         def netzke_ext_css_include(params)
           # ExtJS base
-          res = ["#{Netzke::Core.ext_uri}/resources/css/ext-#{params[:theme]}.css"]
+          res = ["#{Netzke::Core.ext_uri}/resources/ext-theme-#{params[:theme]}/ext-theme-#{params[:theme]}-all.css"]
 
           # Netzke-related dynamic css
-          res << netzke_path(:ext)
+          res << netzke_ext_path
 
           res += Netzke::Core.external_ext_css
 
@@ -28,7 +28,7 @@ module Netzke
           res << "#{Netzke::Core.ext_uri}/locale/ext-lang-#{I18n.locale}" if I18n.locale != :en
 
           # Netzke-related dynamic JavaScript
-          res << netzke_path(:ext)
+          res << netzke_ext_path
 
           javascript_include_tag(*res)
         end