+++ lib/netzke/core/services.rb
@@ -103,11 +103,11 @@ module Netzke::Core
     # +endpoint+ may contain the path to the endpoint in a component down the hierarchy, e.g.:
     #
     #     invoke_endpoint(:users__center__get_data, params)
-    def invoke_endpoint(endpoint, params)
-      if self.class.endpoints[endpoint.to_sym]
-        endpoint_response = Netzke::Core::EndpointResponse.new
-        send("#{endpoint}_endpoint", params, endpoint_response)
+    def invoke_endpoint(endpoint, params, configs = [])
+      endpoint_response = Netzke::Core::EndpointResponse.new
 
+      if has_endpoint?(endpoint)
+        send("#{endpoint}_endpoint", params, endpoint_response)
         endpoint_response
       else
         # Let's try to find it recursively in a component down the hierarchy
@@ -118,12 +118,26 @@ module Netzke::Core
         raise RuntimeError, "Component '#{self.class.name}' does not have endpoint '#{endpoint}'" if !action
 
         if components[child_component]
-          component_instance(child_component).invoke_endpoint(action, params)
+          client_config = configs.shift || {}
+          js_id = client_config.delete("id")
+          cmp_strong_config = {client_config: client_config, js_id: js_id}
+          component_instance(child_component, cmp_strong_config).invoke_endpoint(action, params, configs)
         else
           # component_missing can be overridden if necessary
-          component_missing(child_component)
+          component_missing(child_component, params, endpoint_response)
+          endpoint_response
         end
       end
+    end
+
+    def has_endpoint?(endpoint)
+      !!self.class.endpoints[endpoint.to_sym]
+    end
+
+    # Called when the method_missing tries to processes a non-existing component. Override when needed.
+    # Note: this should actually never happen unless you mess up with Netzke component loading mechanisms.
+    def component_missing(missing_component, params, this)
+      this.netzke_feedback "Unknown component #{missing_component} for #{name}"
     end
 
   end